docs: inline props docs for Dialog
