+++ src/dialog/button.tsx
@@ -1,26 +1,6 @@
-import React, { forwardRef } from 'react';
+import { forwardRef } from 'react';
 
-import { Button as ButtonBase } from '../button';
-import { buttonStyle } from './styles';
-
-type ButtonProps = { autoFocus?: boolean } & React.ComponentProps<
-  typeof ButtonBase
->;
-
-// Override autoFocus behavior to play nicely with react-focus-on
-function ButtonInternal(
-  { autoFocus = false, ...restProps }: ButtonProps,
-  ref: React.Ref<HTMLElement>,
-) {
-  return (
-    <ButtonBase
-      {...restProps}
-      data-autofocus={autoFocus ? true : null}
-      ref={ref}
-      css={buttonStyle()}
-    />
-  );
-}
+import ButtonInternal from './button-internal';
 
 const Button = forwardRef(ButtonInternal);
